// frontend/packages/cozeloop/evaluate-components/src/types/evaluate-target.ts
export type SchemaSourceType = 'set' | 'target';

export type OptionSchema = FieldSchema & {
  schemaSourceType: SchemaSourceType;
};

export interface EvaluatorPro {
  evaluator?: Evaluator;
  evaluatorVersion?: EvaluatorVersion;
  evaluatorVersionDetail?: EvaluatorVersion;
  // key: 评估器字段名，value: 评测目标字段名
  evaluatorMapping?: Record<string, OptionSchema>;
}

export interface CreateExperimentValues extends SubmitExperimentRequest {
  // 评测对象版本详情
  evalTargetVersionDetail?: EvalTargetVersion;

  evaluationSet?: string;
  // 评测集详情
  evaluationSetDetail?: EvaluationSet;
  // 评测集版本
  evaluationSetVersion?: string;
  // 评测集版本详情
  evaluationSetVersionDetail?: EvaluationSetVersion;

  evalTargetType?: EvalTargetType | string | number;
  evalTarget?: string;
  evalTargetVersion?: string;
  // evalTargetVersion?: EvalTargetVersion;
  promptDetail?: prompt.Prompt;
  // key: 评测集字段名，value: 评测目标字段名
  evalTargetMapping?: Record<string, OptionSchema>;

  evaluatorProList?: EvaluatorPro[];
}
// ...
export interface PluginEvalTargetFormProps {
  /**表单数据 */
  formValues: CreateExperimentValues;
  /** 表单数据变化 */
  onChange: (key: keyof CreateExperimentValues, value: unknown) => void;
  /** 创建实验数据, 用于渲染 */
  createExperimentValues: CreateExperimentValues;
  /** 设置创建实验数据 */
  setCreateExperimentValues?: React.Dispatch<
    React.SetStateAction<CreateExperimentValues>
  >;
}

/**
 * 实验创建步骤枚举
 */
export enum ExtCreateStep {
  /** 基础信息 */
  BASE_INFO = 0,
  /** 评测集 */
  EVAL_SET = 1,
  /** 评测对象 */
  EVAL_TARGET = 2,
  /** 评估器 */
  EVALUATOR = 3,
  /** 实验创建 */
  CREATE_EXPERIMENT = 4,
}

/**
 * 评测对象信息
 */
export interface EvalTargetInfo {
  name?: string;
  type?: EvalTargetType;
  color?: string;
  tagColor?: CozeTagColor;
  icon?: React.ReactNode;
}

/** 步骤校验字段, 用于校验表单数据 */
export type ExtraValidFields = {
  [key in ExtCreateStep]?:
    | string[]
    | ((values: CreateExperimentValues) => string[]);
};

export interface EvalTargetDefinition {
  type: EvalTargetType;
  name: string;
  /** 评测对象信息 */
  targetInfo?: EvalTargetInfo;
  /** 评测对象下拉框选择器 */
  selector?: ComponentType<
    SelectProps & {
      /** 选项仅显示名称，不显示头像、描述等复杂信息 */
      onlyShowOptionName?: boolean;
    }
  >;
  /** 评测对象预览器 */
  preview?: ComponentType<{
    /** 评测对象 */
    evalTarget: EvalTarget;
    /** spaceID */
    spaceID: Int64;
    /** 是否显示打开详情链接 */
    enableLinkJump?: boolean;
    /** 尺寸 */
    size?: 'small' | 'medium';
    /** 跳转按钮类名 */
    jumpBtnClassName?: string;
    /** 是否显示图标, 目前 prompt 类型评测对象预览器在实验列表需要显示图标 */
    showIcon?: boolean;
  }>;
  /**
   * 评测对象表单插槽内容
   */
  evalTargetFormSlotContent?: (
    props: PluginEvalTargetFormProps,
  ) => React.JSX.Element | React.ReactNode;
  getEvalTargetVersionOption?: (item: EvalTargetVersion) => {
    value?: string;
    label?: React.ReactNode;
  };
  /**
   * 评测对象预览
   */
  evalTargetView?: (props: {
    /** 渲染数据 */
    values: CreateExperimentValues;
    /** 表单数据 */
    formValues: CreateExperimentValues;
  }) => React.JSX.Element;
  /**
   * 对应表单步骤, 需要额外校验的字段
   */
  extraValidFields?: ExtraValidFields;
  /**
   * 转换创建实验数据
   */
  transformCreateValues?: (
    values: CreateExperimentValues,
  ) => CreateExperimentValues;
}